+++ Lodestone/Modules/Search.php
@@ -2,7 +2,7 @@
 namespace Lodestone\Modules;
 
 use Lodestone\Modules\{
-    Logging\Logger, Routes, Http\HttpRequest
+    Logging\Logger, Routes
 };
 
 trait Search {
